+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 ### Fixed
+- Fixed the logic to reconnect the AppWebsocket automatically in the browser ([#354](https://github.com/holochain/holochain-client-js/pull/354))
 ### Changed
 ### Removed